requirements.txt
# 基础依赖
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
scipy>=1.7.0
tqdm>=4.62.0

# API客户端
requests>=2.26.0
python-dotenv>=0.19.0
openai>=1.0.0
tiktoken>=0.4.0
tenacity>=8.0.0
aiohttp>=3.8.0
httpx>=0.24.0

# 自然语言处理
jieba>=0.42.1
nltk>=3.6.0
rouge-score>=0.0.4
bert-score>=0.3.11
sacrebleu>=2.0.0
py-rouge>=1.1
jieba-fast>=0.53
zhon>=1.1.5

# 图像处理
Pillow>=8.3.0
opencv-python>=4.5.3

# 音频处理
librosa>=0.8.1
soundfile>=0.10.3

# 评测工具
evaluate>=0.4.0
datasets>=2.5.0

# 可视化与报告
seaborn>=0.11.2
plotly>=5.3.0

# 工具
PyYAML>=6.0
jsonlines>=2.0.0
pyarrow>=6.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
